+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added specific value image support
 - Added product delivery time manuell and fixed value
 - Added customer additional text and birthday
+- Added product sort
 - Fixed customer order shipping and billing additional text handling
 - Changed payment pull error tolerance